requirements.txt
django==5.0.3
lxml==5.1.0
djoser==2.2.2
django-cors-headers==4.3.1
djangorestframework==3.15.0                                                               
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
requests==2.31.0
